+++ src/config.spec.ts
@@ -1,0 +1,8 @@
+import {config, internalConfig} from './config';
+
+describe('config', () => {
+    it('changes the default options', () => {
+        config({api: {blacklist: []}});
+        expect(internalConfig).toStrictEqual({api: {blacklist: []}})
+    })
+})