+++ src/index.ts
@@ -1,3 +1,4 @@
 export {register} from './init';
 export {handle} from './handling';
+export {ApiConfigOptions, ConfigOptions, config} from './config';
 export * from './validation';