+++ changelog.md
@@ -3,6 +3,8 @@
 ## v1.0.1
 - Fix changelog
 - Improved readme
+- Added configuration options
+- API Gateway support for more blacklist keys in response
 
 ## v1.0
 - Init project